requirements.txt
feedparser==6.0.11
requests==2.31.0
praw==7.7.1
google-genai>=1.0.0
jinja2==3.1.4
python-dateutil==2.9.0
